server/reviews/serializers.py
from rest_framework import serializers

from reviews.models import Review


class ReviewSerializer(serializers.ModelSerializer):
    class Meta:
        model = Review
        fields = ('id', 'created', 'from_member', 'to_member', 'rating', 'text')
