server/requirements.txt
Django==2.2.2
djangorestframework==3.9.4
django-phonenumber-field==3.0.1
easy-thumbnails==2.6
phonenumberslite==8.10.13
